+++ xenon-common/src/test/java/com/vmware/xenon/common/serialization/TestJsonMapper.java
@@ -13,22 +13,31 @@
 
 package com.vmware.xenon.common.serialization;
 
+import static org.hamcrest.CoreMatchers.containsString;
+import static org.hamcrest.CoreMatchers.not;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertThat;
+import static org.junit.Assert.assertTrue;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.time.Duration;
+import java.util.EnumSet;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.logging.Logger;
+import java.util.regex.Pattern;
 
 import org.junit.Assert;
 import org.junit.Test;
 
+import com.vmware.xenon.common.ServiceDocument;
 import com.vmware.xenon.common.ServiceHost.Arguments;
+import com.vmware.xenon.common.TestGsonConfiguration;
 import com.vmware.xenon.common.Utils;
 import com.vmware.xenon.common.test.TestContext;
 import com.vmware.xenon.services.common.QueryTask.Query;
@@ -109,6 +118,7 @@ public class TestJsonMapper {
         }
     }
 
+    @Test
     public void testPathJsonSerialization() {
         Path p = Paths.get("test");
 
@@ -117,5 +127,40 @@ public class TestJsonMapper {
 
         Arguments arguments = new Arguments();
         Logger.getAnonymousLogger().info(Utils.toJsonHtml(arguments));
+    }
+
+    @Test
+    public void testJsonOptions() {
+        TestGsonConfiguration.AnnotatedDoc testDoc = new TestGsonConfiguration.AnnotatedDoc();
+        testDoc.value = new TestGsonConfiguration.SomeComplexObject("complexA", "complexB");
+        testDoc.sensitivePropertyOptions = "sensitive data1";
+        testDoc.sensitiveUsageOption = "sensitive data2";
+        Pattern containsWhitespacePattern = Pattern.compile("\\s");
+
+        String jsonIncludeSensitiveAndBuiltInPrettyPrinted = Utils.toJson(EnumSet.noneOf(JsonMapper.JsonOptions.class), testDoc);
+        assertThat(jsonIncludeSensitiveAndBuiltInPrettyPrinted, containsString("complexA"));
+        assertThat(jsonIncludeSensitiveAndBuiltInPrettyPrinted, containsString("sensitive"));
+        assertThat(jsonIncludeSensitiveAndBuiltInPrettyPrinted, containsString(ServiceDocument.FIELD_NAME_VERSION));
+        assertTrue("pretty-printed JSON should have whitespaces",
+                containsWhitespacePattern.matcher(jsonIncludeSensitiveAndBuiltInPrettyPrinted).find());
+
+        String jsonExcludeSensitiveIncludesBuiltIn = Utils.toJson(EnumSet.of(JsonMapper.JsonOptions.EXCLUDE_SENSITIVE), testDoc);
+        assertThat(jsonExcludeSensitiveIncludesBuiltIn, containsString("complexA"));
+        assertThat(jsonExcludeSensitiveIncludesBuiltIn, not(containsString("sensitive")));
+        assertThat(jsonExcludeSensitiveIncludesBuiltIn, containsString(ServiceDocument.FIELD_NAME_VERSION));
+
+        String jsonIncludeSensitiveExcludeBuiltIn = Utils.toJson(EnumSet.of(JsonMapper.JsonOptions.EXCLUDE_BUILTIN), testDoc);
+        assertThat(jsonIncludeSensitiveExcludeBuiltIn, containsString("complexA"));
+        assertThat(jsonIncludeSensitiveExcludeBuiltIn, containsString("sensitive"));
+        assertThat(jsonIncludeSensitiveExcludeBuiltIn, not(containsString(ServiceDocument.FIELD_NAME_VERSION)));
+
+        String jsonExcludeSensitiveAndBuiltInCompact = Utils.toJson(
+                EnumSet.of(JsonMapper.JsonOptions.EXCLUDE_SENSITIVE, JsonMapper.JsonOptions.EXCLUDE_BUILTIN, JsonMapper.JsonOptions.COMPACT),
+                testDoc);
+        assertThat(jsonExcludeSensitiveAndBuiltInCompact, containsString("complexA"));
+        assertThat(jsonExcludeSensitiveAndBuiltInCompact, not(containsString("sensitive")));
+        assertThat(jsonExcludeSensitiveAndBuiltInCompact, not(containsString(ServiceDocument.FIELD_NAME_VERSION)));
+        assertFalse("compact JSON should not have whitespaces",
+                containsWhitespacePattern.matcher(jsonExcludeSensitiveAndBuiltInCompact).find());
     }
 }